LayoutTests/http/wpt/web-animations/interfaces/AnimationTimeline/document-timeline-expected.txt


PASS document.timeline identity tests 
FAIL document.timeline.currentTime value tests assert_true: document.timeline.currentTime is positive expected true got false
FAIL document.timeline.currentTime liveness tests assert_greater_than: document.timeline.currentTime increases between script blocks expected a number but got a "object"
FAIL iframe document.timeline.currentTime liveness tests assert_greater_than: iframe document.timeline.currentTime increases between script blocks expected a number but got a "object"
FAIL document.timeline.currentTime time should be the same for all RAF callbacks in a frame assert_greater_than_equal: currentTime should have progressed expected a number but got a "object"
 
